Moved CustomerTableModel to Controller package
